Shorten most FetchSubPhase implementations into lambdas
Most of these aren't configurable in any way. No need to complicate
things by spreading this code out IMO. Using lambda constants we can
shorten this code by quite a bit, remove the weird circular
dependency between the inner hits and the fetch phase and remove some
object instation without state that needlessly complicates things by
obscuring that these things are pure functions.
